move save_feature_map
